tags also deleted when the monitor is deleted

+++ manager/src/main/java/org/dromara/hertzbeat/manager/service/impl/TagServiceImpl.java
@@ -81,7 +81,7 @@ public class TagServiceImpl implements TagService {
     @Override
     public void deleteMonitorSystemTags(Monitor monitor) {
         if (CollectionUtils.isNotEmpty(monitor.getTags())) {
-            List<Tag> tags = monitor.getTags().stream().filter(tag -> tag.getType() == 0).collect(Collectors.toList());
+            List<Tag> tags = monitor.getTags().stream().filter(tag -> tag.getType() == (byte) 0).collect(Collectors.toList());
             tagDao.deleteAll(tags);
         }
     }
